# modules/movies/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "utoipa-axum",
    "description": "Utoipa's axum bindings for seamless integration for the two",
    "contact": {
      "name": "Juha Kukkonen",
      "email": "juha7kukkonen@gmail.com"
    },
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.2.0"
  },
  "paths": {
    "/movie/{id}": {
      "get": {
        "operationId": "movie_details",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaItem"
                }
              }
            }
          }
        }
      }
    },
    "/search": {
      "get": {
        "operationId": "search",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchResult"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/streams/movie/{id}": {
      "get": {
        "operationId": "movie_streams",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Stream"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/streams/start/{info_hash}/{file_idx}": {
      "post": {
        "operationId": "start_stream",
        "parameters": [
          {
            "name": "info_hash",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "file_idx",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartStreamResponse"
                }
              }
            }
          }
        }
      }
    },
    "/streams/tv/{id}/{season}/{episode}": {
      "get": {
        "operationId": "tv_streams",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "season",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "episode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Stream"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/subtitles/cues": {
      "get": {
        "operationId": "subtitle_cues",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "description": "URL of the SRT subtitle file",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubtitleCue"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/subtitles/movie/{id}": {
      "get": {
        "operationId": "movie_subtitles",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubtitleTrack"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/subtitles/tv/{id}/{season}/{episode}": {
      "get": {
        "operationId": "tv_subtitles",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "season",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "episode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubtitleTrack"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tv/{id}": {
      "get": {
        "operationId": "tv_details",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaItem"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Episode": {
        "type": "object",
        "required": [
          "episode_number",
          "name"
        ],
        "properties": {
          "episode_number": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "overview": {
            "type": [
              "string",
              "null"
            ]
          },
          "still_path": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "Genre": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "MediaItem": {
        "type": "object",
        "required": [
          "id",
          "media_type",
          "title",
          "backdrops",
          "genres",
          "videos"
        ],
        "properties": {
          "backdrops": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Genre"
            }
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "imdb_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "logo_path": {
            "type": [
              "string",
              "null"
            ]
          },
          "media_type": {
            "$ref": "#/components/schemas/MediaType"
          },
          "overview": {
            "type": [
              "string",
              "null"
            ]
          },
          "poster_path": {
            "type": [
              "string",
              "null"
            ]
          },
          "rating": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "release_date": {
            "type": [
              "string",
              "null"
            ]
          },
          "runtime": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "seasons": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/Season"
            }
          },
          "tagline": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": "string"
          },
          "videos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Video"
            }
          }
        }
      },
      "MediaType": {
        "type": "string",
        "enum": [
          "movie",
          "tv"
        ]
      },
      "SearchResult": {
        "type": "object",
        "required": [
          "id",
          "media_type",
          "title"
        ],
        "properties": {
          "backdrop_path": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "media_type": {
            "$ref": "#/components/schemas/MediaType"
          },
          "overview": {
            "type": [
              "string",
              "null"
            ]
          },
          "poster_path": {
            "type": [
              "string",
              "null"
            ]
          },
          "release_date": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": "string"
          }
        }
      },
      "Season": {
        "type": "object",
        "required": [
          "id",
          "season_number",
          "name",
          "episode_count",
          "episodes"
        ],
        "properties": {
          "air_date": {
            "type": [
              "string",
              "null"
            ]
          },
          "episode_count": {
            "type": "integer",
            "format": "int64"
          },
          "episodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Episode"
            }
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "poster_path": {
            "type": [
              "string",
              "null"
            ]
          },
          "season_number": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "StartStreamResponse": {
        "type": "object",
        "required": [
          "url"
        ],
        "properties": {
          "url": {
            "type": "string"
          }
        }
      },
      "Stream": {
        "type": "object",
        "required": [
          "info_hash",
          "file_idx",
          "name",
          "title",
          "source",
          "hdr",
          "imax",
          "score"
        ],
        "properties": {
          "audio": {
            "type": [
              "string",
              "null"
            ]
          },
          "codec": {
            "type": [
              "string",
              "null"
            ]
          },
          "file_idx": {
            "type": "integer",
            "format": "int64"
          },
          "hdr": {
            "type": "boolean"
          },
          "imax": {
            "type": "boolean"
          },
          "info_hash": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "resolution": {
            "type": [
              "string",
              "null"
            ]
          },
          "score": {
            "type": "number",
            "format": "double"
          },
          "seeders": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "size_bytes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "size_display": {
            "type": [
              "string",
              "null"
            ]
          },
          "source": {
            "type": "string"
          },
          "source_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": "string"
          }
        }
      },
      "SubtitleCue": {
        "type": "object",
        "required": [
          "start",
          "end",
          "text"
        ],
        "properties": {
          "end": {
            "type": "number",
            "format": "double",
            "description": "End time in seconds"
          },
          "start": {
            "type": "number",
            "format": "double",
            "description": "Start time in seconds"
          },
          "text": {
            "type": "string"
          }
        }
      },
      "SubtitleTrack": {
        "type": "object",
        "required": [
          "id",
          "language",
          "url",
          "score"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "score": {
            "type": "integer",
            "format": "int64",
            "description": "Higher score = more likely to be in sync"
          },
          "url": {
            "type": "string"
          }
        }
      },
      "Video": {
        "type": "object",
        "required": [
          "key",
          "site",
          "name",
          "video_type"
        ],
        "properties": {
          "key": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "site": {
            "type": "string"
          },
          "video_type": {
            "type": "string"
          }
        }
      }
    }
  }
}